+++ Plugins/XmlExport/xmlexport.h
@@ -22,36 +22,36 @@ class XMLEXPORTSHARED_EXPORT XmlExport : public GenericExportPlugin
     public:
         XmlExport();
 
-        QString getFormatName() const;
-        ExportManager::StandardConfigFlags standardOptionsToEnable() const;
-        QString getExportConfigFormName() const;
-        CfgMain* getConfig();
-        void validateOptions();
-        QString defaultFileExtension() const;
+        QString getFormatName() const override;
+        ExportManager::StandardConfigFlags standardOptionsToEnable() const override;
+        QString getExportConfigFormName() const override;
+        CfgMain* getConfig() override;
+        void validateOptions() override;
+        QString defaultFileExtension() const override;
         bool beforeExportQueryResults(const QString& query, QList<QueryExecutor::ResultColumnPtr>& columns,
-                                      const QHash<ExportManager::ExportProviderFlag,QVariant> providedData);
-        bool exportQueryResultsRow(SqlResultsRowPtr row);
-        bool afterExportQueryResults();
-        bool beforeExportSingleTable(const QString& database, const QString& table);
-        bool afterExportSingleTable();
+                                      const QHash<ExportManager::ExportProviderFlag,QVariant> providedData) override;
+        bool exportQueryResultsRow(SqlResultsRowPtr row) override;
+        bool afterExportQueryResults() override;
+        bool beforeExportSingleTable(const QString& database, const QString& table) override;
+        bool afterExportSingleTable() override;
         bool exportTable(const QString& database, const QString& table, const QStringList& columnNames, const QString& ddl, SqliteCreateTablePtr createTable,
-                         const QHash<ExportManager::ExportProviderFlag,QVariant> providedData);
+                         const QHash<ExportManager::ExportProviderFlag,QVariant> providedData) override;
         bool exportVirtualTable(const QString& database, const QString& table, const QStringList& columnNames, const QString& ddl, SqliteCreateVirtualTablePtr createTable,
-                                const QHash<ExportManager::ExportProviderFlag,QVariant> providedData);
-        bool exportTableRow(SqlResultsRowPtr data);
-        bool afterExportTable();
-        bool beforeExportDatabase(const QString& database);
-        bool exportIndex(const QString& database, const QString& name, const QString& ddl, SqliteCreateIndexPtr createIndex);
-        bool exportTrigger(const QString& database, const QString& name, const QString& ddl, SqliteCreateTriggerPtr createTrigger);
+                                const QHash<ExportManager::ExportProviderFlag,QVariant> providedData) override;
+        bool exportTableRow(SqlResultsRowPtr data) override;
+        bool afterExportTable() override;
+        bool beforeExportDatabase(const QString& database) override;
+        bool exportIndex(const QString& database, const QString& name, const QString& ddl, SqliteCreateIndexPtr createIndex) override;
+        bool exportTrigger(const QString& database, const QString& name, const QString& ddl, SqliteCreateTriggerPtr createTrigger) override;
         bool exportView(const QString& database, const QString& name, const QStringList& columnNames, const QString& ddl,
                         SqliteCreateViewPtr createView, const QHash<ExportManager::ExportProviderFlag,QVariant> providedData) override;
-        bool exportViewRow(SqlResultsRowPtr data);
-        bool beforeExportSingleView(const QString& database, const QString& name);
-        bool afterExportSingleView();
-        bool afterExportView();
-        bool afterExportDatabase();
-        bool init();
-        void deinit();
+        bool exportViewRow(SqlResultsRowPtr data) override;
+        bool beforeExportSingleView(const QString& database, const QString& name) override;
+        bool afterExportSingleView() override;
+        bool afterExportView() override;
+        bool afterExportDatabase() override;
+        bool init() override;
+        void deinit() override;
 
     private:
         void setupConfig();