+++ ChangeLog.md
@@ -63,6 +63,7 @@
 - CHANGE: #4475 Using RETURNING clause for INSERTing data to optimize the operation.
 - CHANGE: #3539 File extension in Export Dialog file input is now updated whenever export format is changed.
 - CHANGE: #4688 Added metadata to Windows executables.
+- CHANGE: #2956 Export Dialog respects hotkeys for changing database in the dropdown, using the same hotkey as SQL Editor.
 - CHANGE: Windows x86 (32-bit) binaries will no longer by provided in official builds.
 - BUGFIX: #3079 Enterprise Formatter now handles comments much better than before.
 - BUGFIX: #4986 Unique indexes are allowed for expressions, as they should be.